charnames.t: Fix to work with new charnames.pm
Version 1.33 of charnames.pm is changing in conjunction with core Perl
to make \N{ANY UNKNOWN CHARACTER} fatal, instead of merely being a
warning.  This will now conform to the behavior that Unicode recommends.

This commit causes charnames.t to expect a warning from versions of
charnames.pm earlier than 1.33, and a fatal error beginning with 1.33.